refactor API for huffman encode/decode to allow switching tables easier
Summary: before this diff the API based on simple functions that take HPACK::MessageType as an argument and decide internally which tables to switch. This makes hard to choose a different huffman table, so now we moved all the (static) functions as methods in a HuffTree and inject that in the HPACKEncode/DecodeBuffer instead of the MessageType.

Test Plan: unit tests

Reviewed By: afrind@fb.com

Subscribers: trunkagent, doug, bmatheny, cgheorghe

FB internal diff: D1675948

Tasks: 5585851

Signature: t1:1675948:1415821650:80b3431baaa6eddcaae7846449f0b85da444d085