Don't do 'delete this' in HTTPTransaction
Summary: 'delete this' is a risky pattern. A safer pattern is to use vector/map as one composes lifteimes. Since HTTPSession always outlives all the transactions it comprises, this is a safe change. We have a check in the destructor of HTTPSession already to make sure this vector is empty when the class is destroyed as well.

Test Plan: unit tests with ASAN enabled, will canary

Reviewed By: afrind@fb.com

Subscribers: trunkagent, doug, bmatheny

FB internal diff: D1686512

Signature: t1:1686512:1416444507:1653e423eaedc9e8720959b63efda947ef1e3b30