TraceEventObserver refactor
Summary:
All right, this is the TraceEventObserver refactor diff that all of us are excited about.
The design discussion happened a week ago in the group post: https://www.facebook.com/groups/csti.developers/permalink/765817406800210/
Major things that I did in this diff:
- Change TraceEventContext into containing a vector of TraceEventObservers;
- Create java interface for TraceEventObserver and TraceEventContext; Per result of the discussion in the post, events are batched in cpp and only passed back to java when a request finishes.
- Remove RequestStat from ResponseHandler interface and turned it into a TEO;
- Implemented java version of RequestStats;
- Kill ProxygenLogFilter and turned it into a TEO; Ideally we should directly turn CircularLogSink into a TEO but this diff is already too big.
- Cleanup along the way;

Note that I temporarily disabled the trace events in RetryFilter and RedirectFilter. The data is not urgently needed for now and we should treat it differently in the new model. Other than that, all the functionality should remain the same as it is today.

I also need to cleanup RequestStats all the clients in fbcode who use our httpclient and ResponseHandler. But otherwise it should be good to receive review. My apology in advance for the diff size but it is hard to split this out as it involve some interface change. But actually it is less horrible than the number of lines it shows.

Test Plan: Fix all the tests and added the RequestStatsTest in java.

Reviewed By: subodh@fb.com

Subscribers: abirchall, trunkagent, ps, bmatheny, seanc

FB internal diff: D1692041

Tasks: 5285227

Signature: t1:1692041:1419028717:50dbc8e4d42e50700caac0050157151c3fc1a343