Fix HTTPSession use-after-free on shutdown after txn creation
Summary:
We update the txn (and thus the handler's) pause state after
setting the transaction. This could be an issue if the session is forced
closed immediately in the first callback.

This shows the problem listed in the attached task. The test somehow
doesn't crash even with ASAN enabled, but there is a real memory issue
here. `setNewTransactionPauseState()` is checking state on a deleted
`this`. `this` is deleted from inside `txn->setHandler(handler)`.

Not sure what the best fix is here, but I think the API needs to change
slightly to prevent this. Perhaps
`HTTPTransaction::Handler::setTransaction()` should take a boolean
indicating whether it starts with ingress or egress paused.

Test Plan: new unit test, ran with ASAN

Reviewed By: afrind@fb.com

Subscribers: trunkagent, doug, bmatheny

FB internal diff: D1698100

Tasks: 3596574

Signature: t1:1698100:1420484770:bec5080620b26faba7433c6e98cdada949fb821b