Url-specified rate-limiting of proxygen responses
Summary:
Support for egress rate-limiting in HTTPTransaction.

The implementation is to have HTTPTransaction delay sending bytes if
its' over the rate limit.

-------------

For posterity, here's the original summary discussion about the
tradeoffs of various implementations...this is more moot now that
we decided to push the implementation into HTTPTransaction:

This change is tricky for a variety of reasons.  I'll try to summarize
them here to facilitate discussion on whether this implementation is the
right one.

The first question to answer is whether we want to (a) explicitly pause
sending bytes to the client (as if the client were slow), or (b) just
pause the bytes coming in from upstream (and as a side effect this will
pause egress to the client).

(a) seems preferable for a couple reasons.  First, in the future (and
especially when we start using cache leases) we will likely want to
continue pulling bytes in from the upstream as fast as possible, even if
we're going to be sending them to a given client slowly.  And more
generally, "let's slow down sending bytes to the client" is very
conceptually close to "the client wants us to send bytes more slowly"
and so it seems like code-reuse and general cleanliness are more-likely
offshoots of this approach.

However, (a) has a couple of tricky problems.  First, until now, the
only ways that the client-pause callbacks have been invoked
has been from tcp buffers getting full...and similarly, the
client-resume callbacks are invoked when the tcp buffers aren't full.
So even if we manually pause the client egress, it may get unpaused()
automatically.  One idea I had was to have my RateLimitFilter intercept
these client-initiated pause/resume callbacks and only pass along the
resume call if we weren't also paused for rate-limiting purposes.  This
didn't work though, because the response filter chain is hooked up to
the ServerHandler, not the ClientHandler, and thus doesn't actually
receive callbacks about the client asking us to pause.  We may be able
to make this approach work with a bit of a refactoring in which the
response filters receive more of these flow-control callbacks...we'd
have to think about the design though, and it might be a lot of work.

So I ended up going with (b). [Note: this is not the case anymore,
see top of summary]

Test Plan:
Start up a local proxygen server and a local bigcache.

Set up port forwarding from my laptop to my devserver and update
/etc/hosts to point "scontent.xx.fbcdn.net", then hit various urls like

https://scontent.xx.fbcdn.net/hvideo-xpa1/v/t42.1790-2/760687_10100232494366269_657249045_n.mp4?oh=366fea3a14c5dae8a25482097eb22eb9&oe=54AFFBAE&rl_rate=8000
https://scontent.xx.fbcdn.net/hvideo-xpa1/v/t42.1790-2/760687_10100232494366269_657249045_n.mp4?oh=366fea3a14c5dae8a25482097eb22eb9&oe=54AFFBAE&rl_after=5000&rl_rate=8000

and make sure the video loads, and that it's being downloaded at the
expected pace.  Cancel various downloads in the middle and make sure it
doesn't crash.

Make sure that when whether the result comes from cache or not that it
respects the rate limits set.  Watch bigcache logging (v=4) and see that
the get requests to it do slow down when the url specifies a slow
download speed.

Reviewed By: afrind@fb.com

Subscribers: trunkagent, jianx, bmatheny, cgheorghe, ruibalp

FB internal diff: D1780743

Tasks: 5843089

Signature: t1:1780743:1423078861:c5cf46d8e10d9314743d329ae63499a748c56c91