Add extra info to exception in goaway with error status
Summary:
If there is a goaway with a type that is caused
due to a codec error, or not a NO_ERROR type,
we would forward that to the callbacks with the
same error type.

This changes the ProxygenError type for goaways that
return with a special exception, for logging.

Test Plan: Unit tests

Reviewed By: afrind@fb.com

Subscribers: trunkagent, bmatheny, seanc

FB internal diff: D1909536

Tasks: 6463914

Signature: t1:1909536:1426541750:81f945f8eb8983277973f399cfa4917cbee00f38