Correctly serialize multi-values headers with empty values
Summary:
If a header was set multiple times into HTTPHeaders, and one or more of the values had zero length, then GzipHeaderCodec was emitting an illegal header block.  The SPDY spec says that the value cannot start or end with a \0, and all intermediate values must have non-zero length.

It is legal to serialize a header with an empty value, however.  So skip any duplicate headers with zero length, and only append the \0 if the first header had non-zero length, otherwise overwrite it.

Test Plan: New unit test

Reviewed By: viswanath@fb.com

Subscribers: viswanath, subodh, doug, bmatheny, cgheorghe, seanc, pgriess

FB internal diff: D2108278

Tasks: 7240576