Add padding size to HTTPCodec generateBody/onBody
Summary:
HTTP/2 padding bytes are flow controlled, but both the session and the stream flow control windows are maintained outside of the codec.  So pass the number of flow controlled pad bytes to the callback, and have the window maintainers immediately add them to toAck (since the bytes have been discarded).  Also (primarily for testing) allow padding to be passed to generateBody.

One drawback of the current approach is that we immediately free the padding and add to toAck_, but we cannot generate a window update until the next time an API with a writeBuf is called.  In the pathological case of the test, we don't ack the padding until some read data gets passed to the handler.

Test Plan: New unit test

Reviewed By: xiaosophiawang@fb.com

Subscribers: doug, bmatheny

FB internal diff: D2121370

Tasks: 6902645

Signature: t1:2121370:1433354568:7f35841c0192cdb6f89d5a70a273205ab5add6ef