Implement Rendezvous Hash
Summary:
This is another implementation of Consistent hash ring
that supports weights assigned to sites properly.

** Deploying this change will reshuffle the concurrent bucket assignments **

Several advantages this implementation has over ConsistentHashRing:
- Current implementation of ConsistentHashRing doesn't support
changing weights -- some hash assignments won't be consistent
which results in user shuffling -- simulation shows in the event
of weight change of a cluster, up to 50% of moved users are
unnecessarily shuffled between clusters whose weights are not changed.

- We can patch current implementation to make it support weight changes (D2132617).
However that patch gives build time a runtime complexity of O(W * Ln(W)),
where W is the sum of weights. Once we increase the total weight to the order of
billions, it would take the ring hours to build.

A disadvantage of this approach:

- The complexity of "Get" function is O(N), where N is the number of frontend
clusters.

- The function call to 'pow' is expensive.

The combination of the two function calls above makes the "get" 7x slower than
the current implementation for current weight. It is about 3 micro seconds per
call so I guess it's overhead is trivial -- It also becomes faster than D2132617
once we set the weights to millions.

Test Plan:
Plug this into ConsistentHash inspector(need to change
several lines since the api is slightly different). Notice that
the shuffle rate is consistent to be 0.

Reviewed By: dmitrip@fb.com

Subscribers: praveen, trunkagent, praveenr, frumious, markdrayton, jeffliu, bmatheny, hoo, yunjin

FB internal diff: D2151369

Tasks: 7240485

Signature: t1:2151369:1434585089:aeeae720854700b73fbbe4bb095c2632d95d72ed