HTTPTransaction::sendBody pauses handler immediately when deferredEgress exceeds send window
Summary: This behavior changed a few months back and to support a different scheme for egress flow control, but it's actually a really good signal to pause the handler when the window is full.  It keeps the handler for piling up egress in the HTTPTransaction.

Test Plan: Unit tests

Reviewed By: noamler@fb.com

Subscribers: trunkagent, bmatheny

FB internal diff: D2169419

Tasks: 6607709