Set HTTP2Priority on HTTPMessage
Summary:
This is so the HTTP2Codec can communicate the parsed priority to the CodecCallback, and the caller can specify one in the message.  I didn't want to create a dep from HTTPMessage to HTTP2Framer (where PriorityUpdate is defined), so just made this a std::tuple.

I also added a MAX_STREAM_ID to HTTPCodec that reflects the max for both SPDY and HTTP/2.  The plan is to use a specially configured HTTP2PriorityQueue for both SPDY and HTTP/2 that will rely on virtual streams higher than the max ID.

Test Plan: This could use more tests.

Reviewed By: noamler@fb.com

Subscribers: noamler, bmatheny

FB internal diff: D2196145