Logged flow control hit into network status monitor
Summary: As discussed, we want the data for how frequent we hit the write flow control window full case. This diff added an aggregated number of flow control window closing case of most recent 5 secs. Idealy we want that number for requests(connections) that we care about. But a global aggregated number is easier to implement and should provide strong enough signal for comparison purpose.

Reviewed By: benyluo

Differential Revision: D2478177

fb-gh-sync-id: 50fc54d39f27315c9b9193a05691abdb98e5188a