Handle HTTPSession::readEOF with session/streams blocked on flow contrl
Summary: It was possible to zombie an HTTPSession that had some flow control blockage by sending an abrupt EOF.  To fix I detect the condition and convert the read shutdown into a read+write shutdown.  Note in the per-stream case, we also have to sendAbort to generate some egress that will determine if the socket is also shutdown for writes.

Reviewed By: russoue

Differential Revision: D2583639

fb-gh-sync-id: 1dbd986c32ebf525bf7f8f77e8946511b07dfa54