Refactor HTTPDownstreamSessionTest
Summary: This was our first attempt at learning gmock and there is so much boilerplate that was never abstracted away and made into an easy-to-use API.  I added a bunch of simplified APIs and helpers and really streamlined the test bodies to just what's required to test.  Hopefully this will make it a easier to write C++ tests.

Reviewed By: rpraveen

Differential Revision: D2591702

fb-gh-sync-id: d0fc7ff90084e85bee2765af65895ae78cb7122b