Resume transactions in priority order, pause inline
Summary: The egress prioritization isn't very fair.  Right now we resume all transactions on the session and let them produce effectively unlimited egress for one loop or until the downstream egress buffer limit is exceeded.  This means that low-pri transactions are egressing too, clogging up the socket for high-pri transactions.  So now we will resume transactions in priority order.  As soon as the egress buffer limit is exceeded, we pause all transactions immediately (do not wait for the end of the loop).  This may resume in a bit more pause/resumes, especially for fast sockets that were clearing the entire egress buffer every loop.  I think we could increase the egress buffer limit to compensate for this (effectively we were ignoring the limit before anyways).

Reviewed By: viswanathgs

Differential Revision: D2173627

fb-gh-sync-id: 62aa36228406a75f6c2ca83806ae3311c14dbfb6