A few adjustments for MSVC
The first is to use `FOLLY_TLS` rather than `__thread`.
The second is to not try to ignore some signals, because those signals don't exist under MSVC.
The final change is to switch `parseHTTPDateTime` to use the standard C++ function `std::get_time` rather than the posix function `strptime` to parse the times. This is because `strptime` is not available under MSVC. The format parameter of `std::get_time` is [defined](http://en.cppreference.com/w/cpp/io/manip/get_time) to be the same as `strptime`.