Wire HTTPCodecFactory into HTTPSessionAcceptor
Summary: HTTPSessionAcceptor should be agnostic to the codec being used, but
right now they are inextricably linked.

This prevents users from creating HTTP servers with codecs other than HTTP1x,
SPDY, and HTTP2.

This diff does not actually update HTTPServer to pass a codec factory to its
acceptor, but that is in a follow up diff.

Reviewed By: bugok

Differential Revision: D2693419

fb-gh-sync-id: d1ca92cebdad645406775e9d1160c3a7ca94864c