Clean up HTTP2PriorityQueue
Summary:
This class got completely implemented in the .h file because I was lazy.  Move the more complex functions into the cpp.  Removed some unused code, made some functions private that were not intended to be part of the interface.

This diff is probably impossible to read but it's just rearranging the deck charirs.  No functional changes.

Reviewed By: maxgeorg

Differential Revision: D2722563

fb-gh-sync-id: b4cbc26ffb30324082c1486ab6e0a11399f5ac79