HTTP2PriorityQueue performance fixes
Summary:
There were two perf problems:

1) Calculating the total weight of nodes with egress (including descendants) was done in O(N) time on every call to nextEgress().  This was for simplicity but the perf is terrible, especially for large N with sparse egress nodes.  Instead, update this whenever state changes (including bubbling up).  To ensure correctness, debug builds still invoke the O(N) function and perform CHECKs to make sure we didn't screw this up since it is non-trivial.

2) nextEgress was visiting all nodes of the tree not only nodes with egress.  Maintain a skip list of nodes that have egress (including descendants) and only touch these nodes from nextEgress.

Notes:

a) This code is much more complicated than I ever intended it to be.  I'm not sure if some additional abstraction would help or not, but the code was a little hairy before and now it is a hot mess.  I had to bend over backwards in a couple places so my hyper aggressive CHECKs would pass.

b) The new code could probably be optimized for fewer allocs.  I considered having only a single list of children, partitioned into enqueued/not enqueued halves.  The problem with this is it loses the very nice property that the ordering of transactions is stable.  Maybe replacing the enqueuedChildren_ list iwith something intrusive would help.

Reviewed By: xiaosophiawang

Differential Revision: D2722565

fb-gh-sync-id: a8edba96b22c1d137bdb3e2fc206c3a8929be6f5