Fix two HTTP2PriorityQueue crashes
Summary: It turns out the fragile code to keep the enqueuedIter_ up to date was...well..fragile.  There was a clearly missed else case in addChildren to set the enqueuedIter_ to the end of its new parent's list.  The other bug was more subtle and happened from removeFromTree.  I fixed it completely dequeueing the subtree of the removed node before calling addChildren.

Reviewed By: maxgeorg

Differential Revision: D2793760

fb-gh-sync-id: b59cb51fdc79ac56ed3863a27404aa8e4019114f