Moar fixes for HTTP2PriorityQueue
Summary:
Node::emplaceNode was buggy when the exclusive flag was present, because it left the tree in an inconsistent state before calling addChildren.  Made the state consistent there (and also moved addChildren to the end, so that node->parent_ is correctly set before invoking).

Added a safety statement to detachChild which nulls out parent_ of the detached node.

Node::reparent also had some issues with getting the states correct.  The new code there does not re-use signal/clearPendingEgress, but it has fewer lines and branches.

Lastly propagatePendingEgressClear had a subtle bug where it assumed that the current not was not itself enqueued_, but that function is now called in some places where that may not be true.

Reviewed By: maxgeorg

Differential Revision: D2800983

fb-gh-sync-id: e62d3123851e2f246fb6841f467f66e94b63c65c