Fix caseSensitive flag in HTTPMessage::checkForHeaderToken
Summary:
To the caseSensitive=false, you have to pass a lower case token as well.  This isn't obvious, and makes it hard to pass something defined as a constant with some casing (eg: HTTP2-Settings).

Also, this code did a lot of temporary string allocations for case insensitive searches. Uses a static thread local for the lower case memory.

Reviewed By: russoue

Differential Revision: D2918781

fb-gh-sync-id: 0b5a3d9bfa00e3ee83089f2c8f09e4b9ceb5c0d6
shipit-source-id: 0b5a3d9bfa00e3ee83089f2c8f09e4b9ceb5c0d6