Support EOM in headers-only request in HTTPSession
Summary: SPDY 3 requires headers-only requests to set FLAG_FIN in the SYN_STREAM. Our codec already supports this. We just never utilize it, which turns out to be a bug when talking to some non-FB servers (e.g., twitter). This diff add eom flag to sendHeaders API to allow clients to explicitly set it, so that the codec will handle this properly. It also adds a new rule to the state machine to allow it move from HeadersSend directly to SendingDone, with a eomFlushed event. In HTTPTransaction::sendHeaders() function, if eom is set to true, the State Machine will jump two steps from Start -> HeadersSent -> SendingDone.

Reviewed By: afrind

Differential Revision: D2870238

fb-gh-sync-id: 5f601fb5c0b0fe8c8854f846da650365e2b939ed
shipit-source-id: 5f601fb5c0b0fe8c8854f846da650365e2b939ed