Handle non-default flow control settings in plaintext upgrades
Summary: I realized I missed this test case and the code was incorrect.  The session flow control filter was never sized to the target capacity, and HTTPTransaction::reset() also needs to generate a window update for stream recv window - init window for the upgrade txn.  This is more important for upstream (client) upgrade than downstream, because the server txn upgrades after the POST is complete, so wasn't subject to txn flow control at all.  In fact, I wonder if we should omit the window update from HTTPTransaction::flushWindowUpdate when ingress is complete.

Reviewed By: russoue

Differential Revision: D2987399

fb-gh-sync-id: 27cab84de02adade45dfd119f7eee100eb4ce76d
shipit-source-id: 27cab84de02adade45dfd119f7eee100eb4ce76d