Change generated header codes to work with more headers
Summary:We want to make sure that all the headers will fit in the array. The
old code checked that MAX_HASH_VAL < 256 which isn't wrong, but constricts the
number of headers we can define. We're definitely fine up through 256 headers,
but after that we don't really want to bump up to an array of 65536 strings.

I changed the script that generates this code to define NUM_HEADER_CODES and we
can use that to decide how big the array should be. Then if we overflow
HTTPHeaderCode to uint16_t, we don't need to change anything here. I only sort
of follow the awk script, but reading the generated file confirms it works.

Reviewed By: afrind

Differential Revision: D3165878

fb-gh-sync-id: df1965f176e5cb2b28a50f0e2e5abc124905e1b1
fbshipit-source-id: df1965f176e5cb2b28a50f0e2e5abc124905e1b1