Allow setCapacity to increase the header table size
Summary: If setCapcitity was called with a larger size, it would update the capacity_, but not resize the table.  This could lead to cases where the table *should* have evicted items on add, but did not, and they were later overwritten, and the table became corrupt.  Instead, handle increasing capacity.

Reviewed By: russoue

Differential Revision: D3498979

fbshipit-source-id: ab0e25989cedbf13eca7bdf128d1936917c85cd5