Enforce content-length matches ingress body length
Summary:
This is a must in the HTTP/2 rfc.  It shouldn't affect HTTP/1.1 because "chunked" messages are excluded, and non-chunked HTTP/1.1 messages have the EOM determined by the content-length header in the parser.

Note there's some funkyness around HEAD requests and other responses which are specifically allowed by the RFC to have 0 body and non-zero content-length.

Reviewed By: afrind

Differential Revision: D3541173

fbshipit-source-id: 69b9fd63ef3a687add194e5720a02caf7c1f5641