Prevent duplicates in the priority tree
Summary: The HTTP2PriorityTree makes an assumption that there are no duplicates in the tree.  Look for existing node in addTransaction and update it if present.  This will also CHECK fail if the existing node already had a transaction.  So this will enable PRIORTY frames to precede HEADER frames for the same ID correctly.

Reviewed By: maxgeorg

Differential Revision: D3861913

fbshipit-source-id: 0a81a2d66c32d0068197c9767d735266c6fd6b5f