Add an O(1) index to HTTP2PriorityQueue
Summary:
I noticed now that we call find() in addTransaction, it is taking a lot of CPU.  This is because it is O(N) and we maintain up to 200 virtual nodes in the tree.  Bite the bullet and add a hashtable for quick lookups.  I used boost::intrusive::unordered_set to avoid any extra allocations for this index, though the size of the queue itself is now at least 800 bytes larger to accommodate the buckets.

In the process I discovered that depth calculations were wrong (we were frequently reporting the depth of our parent, rather than the actual depth).  Now children of root should have depth 1, and so on.  Depth calculations now have the possibility of being O(N) in the case where the queue is really a list.  I will recheck the performance profile in production and possible store depth in each node.

Reviewed By: maxgeorg

Differential Revision: D3987961

fbshipit-source-id: edd30ebe4e922abab9c152b980c7b0b3d02380d9