Set lastStreamID to 1 after plaintext upgrade
Summary: The first transaction on a downstream upgraded parallel codec didn't go through the normal path, so lastStreamID_ is still 0.  If the server tries to send a GOAWAY before this request completes, it will incorrectly indicate the request was not processed.  This will cause the request to fail in some way, and may cause the client to erroneously resend the request.

Reviewed By: lnicco

Differential Revision: D4045275

fbshipit-source-id: f7646824a0ccc519ba0d4b3ac07ea1a0fd8b6014