HTTPSession/HTTPTransaction: cast before take negative
Summary: When we have an unsigned int and negate it before pass to a signed int, it doesn't become negative, it just becomes a different positive int. That leads to wrong account of pendingWriteSize_ or pendingWriteSizeDelta_ in HTTPSession, which makes some transaction very almost impossible to resume its egress before it times out itself.

Reviewed By: siyengar

Differential Revision: D4052275

fbshipit-source-id: c00320d34145613076e8a5de81e3308f3753d895