Support HTTP2-Settings for client plaintext upgrade
Summary:
The spec requires that a client send a HTTP2-Settings header with base64url encoded SETTINGS frame payload.  Add a helper method HTTPCodec::requestUpgrade that does this (with the codec default settings).

It turns out to be tricky to do non-default settings, because the settings are owned by the codec, and the codec gets created after a successful upgrade in the current model.  We would have to keep the codec around somewhere during the upgrade request and then switch to it.

Reviewed By: russoue

Differential Revision: D2927926

fbshipit-source-id: d945a5a499913317267f865f291fb2dfce26a656