Fix a case that was eating \r
Summary:
This parser is trying to be very generous regarding \r's, and it makes the code complicated and bug prone.  Normally the end of a field should be terminated with \r\n<boundary>, but this code attempts to allow \n<boundary>.  The problem is when we encounter a \r that does not precede a boundary, we need to make sure it ends up in the result.  The test that was failing was when the \r preceded a partial match (eg: \r\n<not the boundary).  I fixed up the logic, and added several more unit tests cases.  Coverage build shows that all the lines of readToBounary are covered, for whatever that is worth.

The original test case uses random data.  Every so often it would insert a \r\n in the body and fail.

Reviewed By: russoue

Differential Revision: D4246123

fbshipit-source-id: 49ada3eafc7781a6009115a1f6463859ff868848