Further fixes to Travis / Dockerfile
Summary:
For some reason, adding the matrix in D4356576 made Travis emit the wrong shell commands. Compare:
 - https://travis-ci.org/facebook/proxygen/jobs/185714034/config
 - https://travis-ci.org/facebook/proxygen/jobs/185661448/config

So here, I try to work around this by just installing Python via a script.

Also, I noticed that the Dockerfile was not guaranteed to build the same sha1 of Proxygen that was triggering the build, which would result in confusing debugging sessions since ongoing builds would get broken by changes that landed 15-30 minutes later. The fix is, of course, to pass the SHA1 into the Dockerfile.

I'm getting rid of the `BOGUS_JOB` exclude on the theory that maybe we don't need it? E.g. https://github.com/gappleto97/CITest/blob/master/.travis.yml

Reviewed By: lanxi

Differential Revision: D4361706

fbshipit-source-id: a18969c289c6823de0ad3c60c0ce5c9bb5fbfa27