Add ubuntu:8.6 build on Travis
Summary:
The Proxygen tests were segfaulting on Ubuntu 14.04 + gcc 4.9. We're linking against Boost that was built with GCC 4.8. I got concerned if maybe the segfault is due to subtle ABI incompatibilities between 4.8 and 4.9.

To rule this out, I needed a system built with 4.9. Luckily Debian stable (jessie 8.6) provides such a system. Let's add it to the Travis build matrix.

Reviewed By: agoder

Differential Revision: D4364721

fbshipit-source-id: bd847a33c24b6bec81de7bbbb9b9d520f2bb947d