HTTPUpstreamSession support attach/detachEventBase which only detach eventBase and internal objects using it
Summary:
Some of the threadLocals are optional, such as SSLContext, however, attachThreadLocals requires that parameter
In ServiceRouter implementation, a global SSLContext is used, also, thrift API doesn't take in a SSLContext when switching thread to serve the HTTPSession.
To make this happen in thrift, another API is needed, only detach/attach eventBase.

This diff works together with : D4904723

Reviewed By: siyengar

Differential Revision: D4904700

fbshipit-source-id: e284e021055fa2a92b125f888495f574ae402062