Resume pipelined requests as soon as previous response is flushed
Summary:
The code was not resuming a pending pipeline request until after the transaction was destroyed, which can be delayed by byte event tracking such as ack timing.  Further, 4.6 kernel mechanisms for ack timing require that the read callback be installed, so the code could potentially stall a *non* pipelined request when the ack was delayed or reordered.

The fix is to use the count of incomingStreams_ to determine if pipelining is happening, rather than transactions_.size().  This also means we need to resume the transactions following decrementTransactionCount().

Reviewed By: maxgeorg

Differential Revision: D4918637

fbshipit-source-id: c46a322ec9005677ba486c5efcf2b6588c300aa8