Fix pipelining logic check
Summary: There can now be more than two transactions in the set when pipelining is detected, if earlier ones are finished but awaiting byte events (writes paused, or ack events, eg).  Find the previous active one by going backwards from the end instead of assuming it was the first one.

Reviewed By: lnicco

Differential Revision: D5071347

fbshipit-source-id: 6a212f8e1bbbe4dd8209d52779e64c8bef452224