Fix ZlibStreamCompressor
Summary:
ZlibStreamCompressor claimed to support chained IOBufs but actually didn't, it would
loop until it crashes trying to finalize the stream. This diff rewrites the logic to
be simpler and also not manually try to handled the chaining (instead iterating over
each individual buffer). I also stops using Appender so we won't switch over to a new
IOBuf immediately when it becomes too short.

This implementation is inspired by folly/Compression.h and is similar (but supports
streaming).

@public

Reviewed By: afrind

Differential Revision: D5182590

fbshipit-source-id: c24db2cc36a5667cd1f58daf539c18d7125b79c5