Generalize map from client priority level to HTTP2 priority to handle arbitrary tree.
Summary:
This patch updates the proxygen virtual stream based priority scheme from the current system which allows only strict priorities to a flexible system allowing an arbitrary tree of virtual streams, subject only to the limitations in the HTTP/2 specification. An arbitrary prioritization scheme can be implemented by constructing virtual streams corresponding to the desired prioritization and attaching new streams as dependancies of the appropriate virtual stream.

Usage:

The user must define a map from an opaque integer priority level to an  HTTP/2 priority corresponding to the virtual stream. This map is  implemented by the user in a class that extends `HTTPUpstreamSession::PriorityMapBuilder`. A shared pointer to this class is passed into the constructor of `HTTPUpstreamSession`. This method will send the virtual streams and return a unique pointer to a class that extends `HTTPUpstreamSession::PriorityAdapter`. This class implements the map between the user defined priority level and the HTTP/2 priority level.

When the session is started, the `createVirtualStreams` method of `PriorityMapBuilder` is called by `HTTPUpstreamSession::startNow`. The returned pointer to the `PriorityAdapter` map is cached in  `HTTPUpstreamSession`. The HTTP/2 priority that should be used for a new stream dependent on a virtual stream corresponding to a given priority level is then retrieved by calling the `HTTPUpstreamSession::getHTTPPriority(uint8_t level)` method.

The prior strict priority system has been left in place for now, but if both maxLevels and priorityMapBuilder are passed into the `HTTPUpstreamSession` constructor, the maxLevels parameter will be ignored.

Facebook section:
This is the open source part of a larger effort to allow the mapping of Tigon priorities to h2. The FB part will make the builder in `HTTPClient` and pass it into the upstream session constructor. Then the request priority will be mapped to h2 priority in `HTTPTransactionAdaptor::setTransaction()`.

Reviewed By: afrind

Differential Revision: D5136079

fbshipit-source-id: a69d087b0fe64348acac2ded3c69d2d3a68b2843