include acceptorSocketOption in IPConfig
Summary:
copied from D5288772;

we (DTI) team are investigating how to set a tos value in egress traffic from proxygen. The first stage is to set a default tos value for all traffic going through encryption proxy which relies on proxygen, and the final stage would be to set different tos value for different kinds of traffic.

This diff is for the first stage and it is an alternative of D4969110, which exposes acceptorFactory. I found that having a acceptorConfiguration in IPconfig (this diff) is a simpler way to achieve the first stage. As of now I can think of different options to set different tos value for different kinds of traffic, which we can discuss later, but regardless we need a way to set a default tos value for all cases.

Reviewed By: maxgeorg

Differential Revision: D5307757

fbshipit-source-id: ee47c072be401693aa610f068d189bfbc93bf38b