Add an option to HTTP2Codec to enable/disable headroom optimization in
Summary:
There are cases that the IOBuf given to the HTTP2Framer wraps a piece of buffer
that's actually read-only. For those cases, the headroom optimization in the
framer will lead to EXC_BAD_ACCESS / KERN_PROTECTION_FAILURE. Add an option in
HTTP2Codec so that we can disable this optimization for some buffers.

Reviewed By: afrind

Differential Revision: D5418302

fbshipit-source-id: 9fda5ae6d98409fa9c18a6157031d984a6d852b3