Don't use previousy parsed cookies
Summary:
Using previously parsed cookies may cause a use-after-free in the following scenario:

getCookie(...)
getHeaders().add(...)
getCookie(...)

In the case where the call to add triggers a resize of the names or values vector in the underlying HTTPHeaders.  For now, disable this.  We will either remove it completely, or cause getting a writable reference to the headers to unparseCookies

Reviewed By: lanxi

Differential Revision: D5851945

fbshipit-source-id: 7dd7af0dce927936a325a8285898b9f41907f21f