use fbstring for HPACKHeader
Summary:
Internally proxygen is compiled for a platform where std::string is really fbstring.  fbstring is 24 bytes and allows for 23 bytes to be stored in-situ (without an extra heap allocation).  GCC-5 std::strings are 32 bytes and allow 15 bytes in-situ.

Because of the way HPACK header tables are allocated, this change increases the memory required by at least 4kb per connection, and more for cases where the name or value is 16-23 bytes.  So force the HPACK table to use the more memory efficient fbstring.

Reviewed By: yfeldblum, dddmello

Differential Revision: D5808250

fbshipit-source-id: 2a8960822984484f227918ed819d6ca64ed60b65