Add QCRAM support to HPACK, part 1 - seqns and epochs
Summary:
Splitting this diff into three parts for reviewability.  This is implementing QCRAM support as described in https://tools.ietf.org/html/draft-krasic-quic-qcram-00.

Part 1 adds sequence numbers to the encoded blocks and epochs to the table state.  The encoder follows the rules in the draft about when an indexed header may be used.  There is an explicit API to increment the encoder's commit epoch (eg: an ack).  The default behavior is to 'auto-commit' every epoch, which assumes in-order delivery.

Reviewed By: dddmello

Differential Revision: D4942974

fbshipit-source-id: 586d3d7b9b3eda36f8ee0d14f66efa968a0407e4